Refactor: Move identity function to separate module
- Created identity.rs module containing identity function and compose
- Simplified main.rs to keep it minimal
- Added .gitignore for Rust build artifacts
- All tests passing in the new module structure
